Splits out API from view component
