+++ frontend/src/App.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { logout, login, whoami, getSession } from "./api/fetch"
 
 class App extends React.Component {
   constructor(props) {
@@ -13,59 +14,6 @@ class App extends React.Component {
     };
   }
 
-  componentDidMount = () => {
-    this.getSession();
-  };
-
-  getCSRF = () => {
-    fetch("/api/csrf/", {
-      credentials: "same-origin",
-    })
-      .then((res) => {
-        let csrfToken = res.headers.get("X-CSRFToken");
-        this.setState({ csrf: csrfToken });
-        console.log(csrfToken);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
-
-  getSession = () => {
-    fetch("/api/session/", {
-      credentials: "same-origin",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.isAuthenticated) {
-          this.setState({ isAuthenticated: true });
-        } else {
-          this.setState({ isAuthenticated: false });
-          this.getCSRF();
-        }
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
-
-  whoami = () => {
-    fetch("/api/whoami/", {
-      headers: {
-        "Content-Type": "application/json",
-      },
-      credentials: "same-origin",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log("You are logged in as: " + data.username);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
-
   handlePasswordChange = (event) => {
     this.setState({ password: event.target.value });
   };
@@ -74,57 +22,8 @@ class App extends React.Component {
     this.setState({ username: event.target.value });
   };
 
-  isResponseOk(response) {
-    if (response.status >= 200 && response.status <= 299) {
-      return response.json();
-    } else {
-      throw Error(response.statusText);
-    }
-  }
-
-  login = (event) => {
-    event.preventDefault();
-    fetch("/api/login/", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "X-CSRFToken": this.state.csrf,
-      },
-      credentials: "same-origin",
-      body: JSON.stringify({
-        username: this.state.username,
-        password: this.state.password,
-      }),
-    })
-      .then(this.isResponseOk)
-      .then((data) => {
-        console.log(data);
-        this.setState({
-          isAuthenticated: true,
-          username: "",
-          password: "",
-          error: "",
-        });
-      })
-      .catch((err) => {
-        console.log(err);
-        this.setState({ error: "Wrong username or password." });
-      });
-  };
-
-  logout = () => {
-    fetch("/api/logout", {
-      credentials: "same-origin",
-    })
-      .then(this.isResponseOk)
-      .then((data) => {
-        console.log(data);
-        this.setState({ isAuthenticated: false });
-        this.getCSRF();
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+  componentDidMount = () => {
+    getSession();
   };
 
   render() {
@@ -134,7 +33,7 @@ class App extends React.Component {
           <h1>React Cookie Auth</h1>
           <br />
           <h2>Login</h2>
-          <form onSubmit={this.login}>
+          <form onSubmit={login}>
             <div className='form-group'>
               <label htmlFor='username'>Username</label>
               <input
@@ -173,10 +72,10 @@ class App extends React.Component {
       <div className='container mt-3'>
         <h1>React Cookie Auth</h1>
         <p>You are logged in!</p>
-        <button className='btn btn-primary mr-2' onClick={this.whoami}>
+        <button className='btn btn-primary mr-2' onClick={whoami}>
           WhoAmI
         </button>
-        <button className='btn btn-danger' onClick={this.logout}>
+        <button className='btn btn-danger' onClick={logout}>
           Log out
         </button>
       </div>
